Added initial version of multilanguage support.
