+++ idd3/rules/pt/__init__.py
@@ -15,31 +15,21 @@
 # You should have received a copy of the GNU General Public License along with
 # this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from idd3.base import *
-
+from idd3.base import Config
+from idd3.rules.pt import transform
+from idd3.rules import universal
 
 config = Config()
-all_transformations = []
-all_rulesets = []
 
+NON_EMITTED_DETS = ('the', 'a', 'an', 'this', 'these', 'that', 'those')
+NON_EMITTED_DETS = ('o', 'a', 'os', 'as', 'um', 'uma', 'uns', 'umas',
+                    'este', 'esta', 'estes', 'estas',
+                    'esse', 'essa', 'esses', 'essas',
+                    'aquele', 'aquela', 'aqueles', 'aquelas')
+RELATIVE_PRONOUNS = ('que', 'onde', 'cujo', 'cuja', 'cujos', 'cujas')
+GERUND_TAGS = ()
 
-def use_language(module):
-    """Configure idd3's global variables (config, all_transformations,
-        and all_rulesets) using those from module.
-    """
-    global config, all_transformations, all_rulesets
+config.from_object(__name__)
 
-    for key, value in module.config.items():
-        config[key] = value
-
-    while len(all_transformations):
-        del all_transformations[0]
-    all_transformations.extend(module.all_transformations)
-
-    while len(all_rulesets):
-        del all_rulesets[0]
-    all_rulesets.extend(module.all_rulesets)
-
-
-from idd3 import rules
-from idd3 import transform
+all_transformations = transform.all_transformations
+all_rulesets = universal.all_rulesets
