+++ idd3/transform.py
@@ -14,9 +14,7 @@
 #
 # You should have received a copy of the GNU General Public License along with
 # this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from __future__ import print_function, unicode_literals, division
-from idd3 import Relation, Transformation
+from idd3.base import Transformation
 
 
 def delete_indices(relations, indices):
@@ -59,442 +57,3 @@ class RemovePunctuation(Transformation):
             if relation.rel == 'p':
                 indices_to_remove.append(i)
         delete_indices(relations, indices_to_remove)
-
-
-class RemoveParataxisFillers(Transformation):
-    """Removes lexical fillers like 'I mean'."""
-
-    def transform(self, relations):
-        for i, relation in enumerate(relations):
-            if relation.rel == 'parataxis':
-                if relation.word == 'mean' \
-                        and len(relation.deps) == 1\
-                        and relations[relation.deps[0]].word == "I":
-                    delete_indices(relations, [i, relation.deps[0]])
-
-
-class RemoveUtteranceInitialConjunction(Transformation):
-    """Removes commonly used utterance-initial conjuntions (and, then)."""
-
-    def transform(self, relations):
-        if relations[1].word.lower() in ('and', 'then'):
-            delete_indices(relations, [1])
-
-
-class JoinNoLonger(Transformation):
-
-    """Joins 'no longer' as a multiword expression."""
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            if relations[i].word == 'no' and relations[i + 1].word == 'longer':
-                relations[i].rel = 'mwe'
-                relations[i + 1].rel = 'neg'
-
-
-class JoinBecauseOf(Transformation):
-
-    """Joins 'because of' as a multiword expression."""
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            if relations[i].word == 'because' and relations[i + 1].word == 'of':
-                relations[i].rel = 'mwe'
-
-
-class JoinUpTo(Transformation):
-
-    """Joins 'up to'."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-        for i in range(len(relations)):
-            if relations[i].word == 'up'\
-                    and i + 1 < len(relations)\
-                    and relations[i + 1].word == 'to':
-                relations[i].word = 'up to'
-                relations[i].tag = 'RB'
-                indices_to_delete.append(i + 1)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class JoinAtAll(Transformation):
-
-    """Joins 'at all'."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-        for i in range(len(relations)):
-            if relations[i].word == 'at'\
-                    and i + 1 < len(relations)\
-                    and relations[i + 1].word == 'all'\
-                    and not relations[i + 1].deps:
-                relations[i].word = 'at all'
-                indices_to_delete.append(i + 1)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class JoinACoupleOf(Transformation):
-
-    """Joins 'a couple of' into a single node."""
-
-    def transform(self, relations):
-        for i, relation in enumerate(relations):
-            if relation.word\
-                    and relation.word.lower() == 'a'\
-                    and i + 2 < len(relations)\
-                    and relations[i + 1].word == 'couple'\
-                    and relations[i + 2].word == 'of':
-
-                k = Relation.get_children_with_dep('adpobj', relations,
-                                                   i + 2)[0]
-
-                relations[i].head = k
-                relations[i].word = 'a couple of'
-
-                relations[k].head = relations[i + 1].head
-                relations[k].rel = relations[i + 1].rel
-
-                for j in range(i + 4, len(relations)):
-                    if relations[j].head == i + 1:
-                        relations[j].head = k
-
-                delete_indices(relations, [i + 1, i + 2])
-
-
-class JoinANumberOf(Transformation):
-
-    """Joins 'a number of' into a single node."""
-
-    def transform(self, relations):
-        for i, relation in enumerate(relations):
-            if relation.word\
-                    and relation.word.lower() == 'a'\
-                    and i + 2 < len(relations)\
-                    and relations[i + 1].word == 'number'\
-                    and relations[i + 2].word == 'of':
-
-                k = Relation.get_children_with_dep('adpobj', relations,
-                                                   i + 2)[0]
-
-                relations[i].head = k
-                relations[i].word = 'a number of'
-
-                relations[k].head = relations[i + 1].head
-                relations[k].rel = relations[i + 1].rel
-
-                for j in range(i + 4, len(relations)):
-                    if relations[j].head == i + 1:
-                        relations[j].head = k
-
-                delete_indices(relations, [i + 1, i + 2])
-
-
-class JoinOneDay(Transformation):
-
-    """Join 'one day' as temporal modifier."""
-
-    def transform(self, relations):
-        for i, relation in enumerate(relations):
-            if relation.word == 'day' and relation.rel == 'nmod'\
-                    and i - 1 >= 0 and relations[i - 1].word.lower() == 'one':
-                relation.word = relations[i - 1].word + ' ' + relation.word
-                delete_indices(relations, [i - 1])
-
-
-class JoinFirstOfAll(Transformation):
-
-    """Joins 'first of all' into a single node."""
-
-    def transform(self, relations):
-        for i, relation in enumerate(relations):
-            if relation.word\
-                    and relation.word.lower() == 'first'\
-                    and i + 2 < len(relations)\
-                    and relations[i + 1].word == 'of'\
-                    and relations[i + 2].word == 'all':
-                relation.word += ' of all'
-                relation.rel = 'advmod'
-                delete_indices(relations, [i + 1, i + 2])
-
-
-class JoinNotEven(Transformation):
-
-    """Handles 'not even', joining it in a single node."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-
-        for i in range(1, len(relations)):
-            if relations[i].word.lower() == 'not'\
-                    and relations[i].tag == 'RB'\
-                    and relations[i].head == i + 1\
-                    and relations[i + 1].word == 'even'\
-                    and relations[i + 1].tag == 'RB':
-                relations[i + 1].word = relations[i].word +\
-                    ' ' + relations[i + 1].word
-                indices_to_delete.append(i)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class JoinMultiWordExpressions(Transformation):
-    """Joins multi-word expressions in a single node. """
-
-    def transform(self, relations):
-        mwes = {}
-
-        for relation in relations:
-            if relation.rel == 'mwe':
-                if relation.head not in mwes:
-                    mwes[relation.head] = [relation.address]
-                else:
-                    mwes[relation.head].append(relation.address)
-
-        for head, deps in mwes.items():
-            new_word = ' '.join([relations[i].word
-                                 for i in sorted(deps + [head])])
-            relations[head].word = new_word
-
-            delete_indices(relations, deps)
-
-
-class JoinPhrasalModifiers(Transformation):
-    """Joins phrasal modal and aspectual markers (e.g., "have to", "ought to",
-        "used to", etc) to the main verb of a sentence. """
-
-    verb_forms = ['have', 'has', 'had', 'ought', 'use', 'uses', 'used']
-
-    def transform(self, relations):
-        for index, relation in enumerate(relations):
-            if relation.rel in ('null', 'ROOT', 'xcomp', 'rcmod')\
-                    and relation.tag in ('VBZ', 'VBD', 'VBP', 'VB')\
-                    and relation.word in self.verb_forms:
-                xcomp_indices = Relation.get_children_with_dep('xcomp',
-                                                               relations,
-                                                               index)
-                if xcomp_indices == []:
-                    continue
-                else:
-                    xcomp_index = xcomp_indices[0]
-
-                if relations[xcomp_index].tag == 'VB':
-                    aux_indices = Relation.get_children_with_dep('aux',
-                                                                 relations,
-                                                                 xcomp_index)
-                    to_index = [index for index in aux_indices
-                                if relations[index].tag == 'TO'][0]
-
-                    # Append 'to' and the xcomp head to main verb.
-                    relations[index].word += ' to ' + \
-                        relations[xcomp_index].word
-
-                    # Remove 'aux' and 'xcomp' relations.
-                    delete_indices(relations, [to_index, xcomp_index])
-
-                    # Change head of relations pointing to xcomp to point to
-                    #   the main verb.
-                    for i, rel in enumerate(relations):
-                        if rel.head == xcomp_index - 2:
-                            rel.head = index
-                            relations[index].deps.append(i)
-
-                    relations[index].deps.sort()
-
-
-class JoinDoublePrepositions(Transformation):
-
-    """Handles double prepositions, as in 'As of 2014'."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-
-        for i in range(len(relations)):
-            if relations[i].tag == "IN":
-                if i + 1 < len(relations) and relations[i + 1].tag == "IN":
-                    relations[i].word += ' ' + relations[i + 1].word
-
-                    for j in range(i + 1, len(relations)):
-                        if relations[j].head == i + 1:
-                            relations[j].head = i
-                            relations[i].deps.append(j)
-
-                    relations[i].deps.sort()
-                    indices_to_delete.append(i + 1)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class JoinExpletives(Transformation):
-
-    """Handles expletives, as in 'there is', or 'there are'."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-
-        for i in range(len(relations)):
-            if relations[i].tag == "EX":
-                relations[i + 1].word = relations[i].word +\
-                    ' ' + relations[i + 1].word
-                indices_to_delete.append(i)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class JoinThereModifiers(Transformation):
-
-    """Handles phrases like 'up there', 'down there', and 'right there'."""
-
-    def transform(self, relations):
-        indices_to_delete = []
-
-        for i in range(1, len(relations)):
-            if relations[i].word.lower() in ('up', 'down', 'right')\
-                    and i + 1 < len(relations)\
-                    and relations[i + 1].word == 'there':
-                relations[i + 1].word = relations[i].word +\
-                    ' ' + relations[i + 1].word
-                indices_to_delete.append(i)
-
-        delete_indices(relations, indices_to_delete)
-
-
-class FixAdjectiveRepetition(Transformation):
-
-    """Handles adjective repetition as intensifier (e.g., she was gone a long
-        long time.). In this case, the first adjective is turned into an
-        adverb and connected to the second adjective."""
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            if relations[i].tag == 'JJ':
-                if i + 1 < len(relations)\
-                        and relations[i + 1].tag == 'JJ'\
-                        and relations[i + 1].word == relations[i].word\
-                        and relations[i + 1].head == relations[i].head\
-                        and relations[i + 1].rel == relations[i].rel:
-                    relations[i].tag = 'RB'
-                    relations[i].head = i + 1
-                    relations[i].rel = 'advmod'
-                    relations[i + 1].deps.append(i)
-
-
-class FixAdverbRepetition(Transformation):
-
-    """Handles adverb repetition as intensifier (e.g., he is very very sick).
-        In this case, we make sure the first adverb is connected to the second,
-        not to the following word (usually an adjective)."""
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            if relations[i].tag == 'RB':
-                if i + 1 < len(relations)\
-                        and relations[i + 1].tag == 'RB'\
-                        and relations[i + 1].word == relations[i].word\
-                        and relations[i + 1].rel == relations[i].rel\
-                        and relations[i].head != i + 1:
-                    relations[i].head = i + 1
-                    relations[i + 1].deps.append(i)
-                    relations[i + 1].deps.sort()
-
-
-class FixReflexivePronouns(Transformation):
-
-    """Handles reflexive pronouns following nouns. Here, we connect the
-        pronoun to the previous noun as an adjectival modifier."""
-
-    reflexive_pronouns = ['myself', 'yourself', 'himself', 'herself', 'itself',
-                          'ourselves', 'yourselves', 'themselves']
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            if relations[i].tag == 'PRP'\
-                    and relations[i].word in self.reflexive_pronouns\
-                    and relations[i - 1].tag in ('NN', 'NNS', 'NNP', 'NNPS'):
-                relations[relations[i].head].deps.remove(i)
-                relations[i].head = i - 1
-                relations[i].rel = 'amod'
-                relations[i - 1].deps.append(i)
-                relations[i - 1].deps.sort()
-
-
-class FixXcompAttributions(Transformation):
-
-    """Turns xcomp relations with no cop children to 'what'."""
-
-# TODO: check if this is still applicable.
-
-    def transform(self, relations):
-        for index, relation in enumerate(relations):
-            if relation.rel == 'xcomp'\
-                    and relation.tag in ('NN', 'NNS', 'NNP', 'NNPS', 'JJ'):
-                if not Relation.get_children_with_dep('cop', relations, index):
-                    relation.rel = 'what'
-
-                    # The subject of this xcomp should become the direct object
-                    #   of the main verb.
-                    xcomp_subj = Relation.get_children_with_dep('nsubj',
-                                                                relations,
-                                                                index)
-                    if xcomp_subj:
-                        relations[xcomp_subj[0]].head = relations[index].head
-                        relations[xcomp_subj[0]].rel = 'dobj'
-                        relations[relations[index].head].deps.append(
-                            relations[xcomp_subj[0]].address)
-                        relations[relations[index].head].deps.sort()
-                        relations[index].deps.remove(relations[xcomp_subj[0]]
-                                                     .address)
-
-
-class TransformCompmodJoin(Transformation):
-
-    """Transforms compmod into compmod-join if both words start with capital
-        letters."""
-
-    def transform(self, relations):
-        for relation in relations:
-            if relation.rel == 'compmod':
-                if relation.word[0].isupper()\
-                        and relations[relation.head].word[0].isupper():
-                    relation.rel = 'compmod-join'
-
-
-class TransformCcIntoPreconj(Transformation):
-
-    """Transforms cc back to preconj, when it's the case."""
-
-    def transform(self, relations):
-        for i in range(len(relations)):
-            cc_indices = Relation.get_children_with_dep('cc', relations, i)
-            conj_indices = Relation.get_children_with_dep('conj', relations, i)
-
-            if cc_indices and not conj_indices:
-                relations[cc_indices[0]].rel = 'preconj'
-
-
-all_transformations = [RemovePunctuation(),
-                       RemoveParataxisFillers(),
-                       RemoveUtteranceInitialConjunction(),
-                       JoinNoLonger(),
-                       JoinBecauseOf(),
-                       JoinUpTo(),
-                       JoinAtAll(),
-                       JoinACoupleOf(),
-                       JoinANumberOf(),
-                       JoinOneDay(),
-                       JoinFirstOfAll(),
-                       JoinNotEven(),
-                       JoinMultiWordExpressions(),
-                       JoinPhrasalModifiers(),
-                       JoinDoublePrepositions(),
-                       JoinExpletives(),
-                       JoinThereModifiers(),
-                       FixAdjectiveRepetition(),
-                       FixAdverbRepetition(),
-                       FixReflexivePronouns(),
-                       FixXcompAttributions(),
-                       TransformCompmodJoin(),
-                       TransformCcIntoPreconj(),
-                       ]
